requirements.txt
Django==1.6
djangorestframework==2.3.13
South==0.8.4